add a util function
